[BP] MB-29531: Replace revSeqno with a 48-bit counter
Prevent a value too large to be stored in couchstore
from being placed into Item/StoredValue and also the
_local document (via vbucket_state).

Reviewed-on: http://review.couchbase.org/93800
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>

+++ tests/ep_testsuite.cc
@@ -3988,7 +3988,7 @@ static enum test_result test_revid(ENGINE_HANDLE *h, ENGINE_HANDLE_V1 *h1)
 
         checkeq(PROTOCOL_BINARY_RESPONSE_SUCCESS, last_status.load(),
                 "Expected success");
-        checkeq(ii, last_meta.revSeqno, "Unexpected sequence number");
+        checkeq(ii, uint64_t(last_meta.revSeqno), "Unexpected sequence number");
     }
 
     return SUCCESS;
@@ -4723,7 +4723,7 @@ static enum test_result test_set_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 0, "Invalid result for flags");
     check(last_meta.exptime == 0, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 1, "Invalid result for seqno");
+    check(last_meta.revSeqno == 1ull, "Invalid result for seqno");
 
     // Check that set with correct cas succeeds
     set_ret_meta(h, h1, "key", 3, "value", 5, 0, last_meta.cas, 10, 1735689600);
@@ -4735,7 +4735,7 @@ static enum test_result test_set_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 10, "Invalid result for flags");
     check(last_meta.exptime == 1735689600, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 2, "Invalid result for seqno");
+    check(last_meta.revSeqno == 2ull, "Invalid result for seqno");
 
     // Check that updating an item with no cas succeeds
     set_ret_meta(h, h1, "key", 3, "value", 5, 0, 0, 5, 0);
@@ -4747,7 +4747,7 @@ static enum test_result test_set_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 5, "Invalid result for flags");
     check(last_meta.exptime == 0, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 3, "Invalid result for seqno");
+    check(last_meta.revSeqno == 3ull, "Invalid result for seqno");
 
     // Check that updating an item with the wrong cas fails
     set_ret_meta(h, h1, "key", 3, "value", 5, 0, last_meta.cas + 1, 5, 0);
@@ -4821,7 +4821,7 @@ static enum test_result test_add_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 0, "Invalid result for flags");
     check(last_meta.exptime == 0, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 1, "Invalid result for seqno");
+    check(last_meta.revSeqno == 1ull, "Invalid result for seqno");
 
     // Check that re-adding a key fails
     add_ret_meta(h, h1, "key", 3, "value", 5, 0, 0, 0, 0);
@@ -4838,7 +4838,7 @@ static enum test_result test_add_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 10, "Invalid result for flags");
     check(last_meta.exptime == 1735689600, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 1, "Invalid result for seqno");
+    check(last_meta.revSeqno == 1ull, "Invalid result for seqno");
 
     return SUCCESS;
 }
@@ -4908,7 +4908,7 @@ static enum test_result test_del_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 0, "Invalid result for flags");
     check(last_meta.exptime == 0, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 1, "Invalid result for seqno");
+    check(last_meta.revSeqno == 1ull, "Invalid result for seqno");
 
     del_ret_meta(h, h1, "key", 3, 0, 0);
     checkeq(PROTOCOL_BINARY_RESPONSE_SUCCESS, last_status.load(),
@@ -4919,7 +4919,7 @@ static enum test_result test_del_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 0, "Invalid result for flags");
     check(last_meta.exptime == 0, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 2, "Invalid result for seqno");
+    check(last_meta.revSeqno == 2ull, "Invalid result for seqno");
 
     // Check that deleting a key with a cas succeeds.
     add_ret_meta(h, h1, "key", 3, "value", 5, 0, 0, 10, 1735689600);
@@ -4929,7 +4929,7 @@ static enum test_result test_del_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 10, "Invalid result for flags");
     check(last_meta.exptime == 1735689600, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 3, "Invalid result for seqno");
+    check(last_meta.revSeqno == 3ull, "Invalid result for seqno");
 
     del_ret_meta(h, h1, "key", 3, 0, last_meta.cas);
     checkeq(PROTOCOL_BINARY_RESPONSE_SUCCESS, last_status.load(),
@@ -4940,7 +4940,7 @@ static enum test_result test_del_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 10, "Invalid result for flags");
     check(last_meta.exptime == 1735689600, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 4, "Invalid result for seqno");
+    check(last_meta.revSeqno == 4ull, "Invalid result for seqno");
 
     // Check that deleting a key with the wrong cas fails
     add_ret_meta(h, h1, "key", 3, "value", 5, 0, 0, 0, 0);
@@ -4950,7 +4950,7 @@ static enum test_result test_del_ret_meta(ENGINE_HANDLE *h,
     check(last_meta.flags == 0, "Invalid result for flags");
     check(last_meta.exptime == 0, "Invalid result for expiration");
     check(last_meta.cas != 0, "Invalid result for cas");
-    check(last_meta.revSeqno == 5, "Invalid result for seqno");
+    check(last_meta.revSeqno == 5ull, "Invalid result for seqno");
 
     del_ret_meta(h, h1, "key", 3, 0, last_meta.cas + 1);
     checkeq(PROTOCOL_BINARY_RESPONSE_KEY_EEXISTS, last_status.load(),
